feat(work-items): support fractional story points
Change story_points from integer to numeric(6,2) so plan estimates can be
fractional per SRS §9.4, mirroring the numeric-as-string pattern used for
task hours.

- work_items.story_points -> numeric(6,2); domain/DTO types number->string
  with 2dp normalisation on input and number output at the API boundary
- align iteration_daily_snapshots point columns (total/completed/remaining)
  from integer to numeric(8,2), matching release_daily_snapshots
- iteration-status add-item + list plan estimate made fractional-consistent
- reporting/aggregation queries cast points via ::float8 / ::numeric to stay
  precision-safe; seed data updated
- migration 0044_story_points_decimal (lossless int->numeric widening)

+++ apps/worker/src/cron/snapshot.cron.ts
@@ -106,8 +106,8 @@ export class SnapshotCronService {
       .select({
         totalItems: sql<number>`count(*)::int`,
         completedItems: sql<number>`count(*) filter (where ${workflowStatuses.category} = 'done')::int`,
-        totalPoints: sql<number>`coalesce(sum(${workItems.storyPoints}), 0)::int`,
-        completedPoints: sql<number>`coalesce(sum(${workItems.storyPoints}) filter (where ${workflowStatuses.category} = 'done'), 0)::int`,
+        totalPoints: sql<number>`coalesce(sum(${workItems.storyPoints}), 0)::float8`,
+        completedPoints: sql<number>`coalesce(sum(${workItems.storyPoints}) filter (where ${workflowStatuses.category} = 'done'), 0)::float8`,
       })
       .from(workItems)
       .innerJoin(
